Move input payloads into peripheral package
