+++ docs/api/input_bus.md
@@ -12,7 +12,7 @@ the shared `StateStore` captures the latest value for each `(event_type, produce
 1. Inject or request an `EventBus` instance. `GameLoop` exposes the bus via
    `GameLoop.event_bus` and propagates it to detected peripherals when
    `ENABLE_INPUT_EVENT_BUS` is enabled.
-1. Compose payloads with the helpers in `heart.events.types`. For example,
+1. Compose payloads with the helpers in `heart.peripheral.input_payloads`. For example,
    `AccelerometerVector(x, y, z).to_input(producer_id=...)` yields a typed input.
 1. Call `event_bus.emit(input_event)` inside the peripheral once the payload is
    validated. Exceptions raised by subscribers are logged and do not block
@@ -56,7 +56,7 @@ another renderer to produce diagrams for incident reports.
 
 - Use the concrete `EventBus` in unit tests and subscribe to the event types
   under test to assert emitted payloads.
-- The helpers in `heart.events.types` return full `Input` instances, enabling
+- The helpers in `heart.peripheral.input_payloads` return full `Input` instances, enabling
   direct comparison in tests without constructing dictionaries manually.
 - `StateStore` objects are safe to share across threads; snapshots return
   read-only `MappingProxyType` views so tests can verify immutability.
