+++ docs/planning/peripheral_color_sync.md
@@ -56,7 +56,7 @@ Implementation proceeds in layers. First, we define the `OutputChannel` protocol
 
 Configuration ties everything together. We extend the runtime configuration schema so that `heart.environment.EnvironmentConfig` accepts an `ambient_channels` section describing each peripheral target, the statistics stream it consumes, and optional transforms. During environment bootstrapping, `Environment._initialize_peripherals` reads this mapping, instantiates an adapter per entry, and registers them with `PeripheralManager`. Developers can then add a single YAML block to map "screen_average" to a desk light, without modifying renderers.
 
-Validation relies on explicit observability. We add structured events under `src/heart/events/types.py` to describe ambient updates (`peripheral.ambient.colour`). The adapter emits these events so downstream tooling—like the debugging CLI in `src/heart/x/cli.py`—can introspect the live colour and verify it matches expectations. A lightweight recorder stored in `src/heart/display/recorder.py` captures frame samples alongside statistics so reviewers can replay the session and confirm the derived colour traces align with the visual scene.
+Validation relies on explicit observability. We add structured events under `src/heart/peripheral/input_payloads.py` to describe ambient updates (`peripheral.ambient.colour`). The adapter emits these events so downstream tooling—like the debugging CLI in `src/heart/x/cli.py`—can introspect the live colour and verify it matches expectations. A lightweight recorder stored in `src/heart/display/recorder.py` captures frame samples alongside statistics so reviewers can replay the session and confirm the derived colour traces align with the visual scene.
 
 Finally, we produce developer tooling to experiment with channel mappings. A new `scripts/ambient_preview.py` command reads the registry output and mirrors the colour in a terminal swatch, mirroring the approach taken by `scripts/render_code_flow.py`. This tightens the edit-test loop for artists who want to tune smoothing parameters or combine multiple statistics (for example, averaging only the top half of the screen for ceiling lights). Packaging these utilities with the plan ensures the workflow remains more expressive than streaming a raw framebuffer.
