--- src/heart/events/README.md
@@ -1,1 +1,0 @@
-Lightweight event definitions used across the Heart runtime. Shared type definitions keep device input, sensor readings, and control signals consistent between modules.
