+++ src/heart/peripheral/led_matrix.py
@@ -8,8 +8,8 @@ from typing import Any, Mapping
 
 from PIL import Image
 
-from heart.events.types import DisplayFrame
 from heart.peripheral.core import Peripheral
+from heart.peripheral.input_payloads import DisplayFrame
 from heart.utilities.logging import get_logger
 
 _LOGGER = get_logger(__name__)
