+++ src/heart/peripheral/microphone.py
@@ -11,8 +11,8 @@ from typing import Any, Self, cast
 
 import numpy as np
 
-from heart.events.types import MicrophoneLevel
 from heart.peripheral.core import Peripheral
+from heart.peripheral.input_payloads import MicrophoneLevel
 from heart.utilities.logging import get_logger
 
 logger = get_logger(__name__)
