+++ src/heart/peripheral/sensor.py
@@ -8,8 +8,8 @@ import reactivex
 import serial
 from reactivex import operators as ops
 
-from heart.events.types import AccelerometerVector
 from heart.peripheral.core import Peripheral, PeripheralInfo, PeripheralTag
+from heart.peripheral.input_payloads import AccelerometerVector
 from heart.utilities.env import get_device_ports
 from heart.utilities.logging import get_logger
